requirements.txt
opencv-python
ultralytics
seaborn
scikit-learn
scipy
matplotlib
pandas
numpy
torch
torchvision
torchaudio
albumentations
pyyaml
tqdm
wandb
tensorboard